fix: updated import path to match renamed component
