+++ src/app/(static)/page.tsx
@@ -11,7 +11,6 @@ import ReviewSection from "@/components/homepage/review-section";
 import NewsSection from "@/components/homepage/news-section";
 import PartnersSection from "@/components/homepage/partners";
 import CampaignList from "@/components/campaign/campaignlist";
-import ShikshaAid from "@/components/aids/shiksha-aid/shikapage";
 
 export default async function HomePage() {
 	try {
@@ -26,7 +25,6 @@ export default async function HomePage() {
 				<ReviewSection />
 				<NewsSection articles={articles} />
 				<AssuranceSection />
-				<ShikshaAid />
 				<PartnersSection />
 			</main>
 		);
