+++ src/app/aids/shiksha/page.tsx
@@ -2,7 +2,7 @@ import CalltoActionSection from "@/components/aids/shiksha-aid/calltoactionsecti
 import Challengessection from "@/components/aids/shiksha-aid/challengessection";
 import Donationsection from "@/components/aids/shiksha-aid/donationplans";
 import Introsection from "@/components/aids/shiksha-aid/introsection";
-import Pillarssection from "@/components/aids/shiksha-aid/pillars_of_program";
+import Pillarssection from "@/components/aids/shiksha-aid/pillarsofprogram";
 import Campaignsection from "@/components/aids/shiksha-aid/recyclecampaign";
 import ShikshaAidPage from "@/components/aids/shiksha-aid/shikapage";
 import Visionsection from "@/components/aids/shiksha-aid/visionsection";
@@ -12,7 +12,6 @@ export default function ShikshaAid() {
 			<ShikshaAidPage />
 			<Introsection />
 			<Pillarssection />
-
 			<Donationsection />
 			<Challengessection />
 			<Campaignsection />
